+++ CHANGELOG.md
@@ -48,6 +48,7 @@
 * Hostname in OSC 7 URI not being validated.
 * OSC 4 with multiple `c;spec` pairs.
 * Alt+Return to emit "ESC \r".
+* Trackpad sloooow scrolling to eventually scroll a line.
 
 
 ### Security